Use Elixir for 2015 Day 08 challenge
I had some trouble working with Elm for this one. The algorithms seemed
to work perfect and translated easily to Elixir code, but Elm was
automatically parsing the escape sequences instead of leaving the raw
string intact. I thought about using `sed` or something to manually edit
the raw input, but I figured I'd try Elixir for this one.