Parsing input for Day 9 graph data
I'm still not sure how to parse strings into types using the Elm parser
(or if that defeats the purpose).

The input is structured nicely with each line in this shape:

Location to Location = Distance

I hard coded the locations into a union type for now since there are
only 8 of them. But I could see Part II having a lot more locations.